move headers and config to devices
