+++ mio/devices/base/__init__.py
@@ -1,0 +1,6 @@
+"""Base device classes"""
+
+from mio.devices.base.device import Device
+from mio.devices.base.headers import BufferHeader
+
+__all__ = ["BufferHeader", "Device"]
