+++ mio/devices/sdcard/__init__.py
@@ -1,0 +1,12 @@
+"""SDCard-based miniscopes, also known as the "wire free" miniscopes"""
+
+from mio.devices.sdcard.device import SDCardDevice
+from mio.devices.sdcard.headers import SDBufferHeader, SDConfig, SDHeaderPositions, SDLayout
+
+__all__ = [
+    "SDBufferHeader",
+    "SDCardDevice",
+    "SDConfig",
+    "SDHeaderPositions",
+    "SDLayout",
+]
