+++ mio/devices/stream/__init__.py
@@ -1,0 +1,19 @@
+"""
+Base classes used by streaming miniscopes,
+like the miniscope zero and MSUS
+"""
+
+# ruff: noqa: I001
+
+from mio.devices.stream.headers import StreamBufferHeader, StreamDevRuntime, StreamPlotterConfig
+from mio.devices.stream.config import StreamDevConfig
+from mio.devices.stream.device import StreamDevice, iter_buffers
+
+__all__ = [
+    "StreamBufferHeader",
+    "StreamDevice",
+    "StreamDevConfig",
+    "StreamDevRuntime",
+    "StreamPlotterConfig",
+    "iter_buffers",
+]
