+++ mio/exceptions.py
@@ -5,14 +5,14 @@ Custom exceptions!
 
 class InvalidSDException(Exception):
     """
-    Raised when :class:`.io.SDCard` is used with a drive that doesn't have the
+    Raised when :class:`.io.SDCardDevice` is used with a drive that doesn't have the
     appropriate WRITE KEYS in its header
     """
 
 
 class EndOfRecordingException(StopIteration):
     """
-    Raised when :class:`.io.SDCard` is at the end of the available recording!
+    Raised when :class:`.io.SDCardDevice` is at the end of the available recording!
     """
