+++ mio/io/__init__.py
@@ -2,13 +2,13 @@
 I/O for files and SDCards
 """
 
+from mio.devices.sdcard.device import SDCardDevice
 from mio.io.csv import BufferedCSVWriter
-from mio.io.sdcard import SDCard
 from mio.io.video import VideoReader, VideoWriter
 
 __all__ = [
     "BufferedCSVWriter",
-    "SDCard",
+    "SDCardDevice",
     "VideoReader",
     "VideoWriter",
 ]
